Split bevy_ggrs example into lib & bin
